refactor(templates): parse template hook
