+++ app/react/portainer/templates/custom-templates/CreateView/InnerForm.tsx
@@ -4,11 +4,7 @@ import { CommonFields } from '@/react/portainer/custom-templates/components/Comm
 import { CustomTemplatesVariablesDefinitionField } from '@/react/portainer/custom-templates/components/CustomTemplatesVariablesDefinitionField';
 import { PlatformField } from '@/react/portainer/custom-templates/components/PlatformSelector';
 import { GitForm } from '@/react/portainer/gitops/GitForm';
-import {
-  getTemplateVariables,
-  intersectVariables,
-  isTemplateVariablesEnabled,
-} from '@/react/portainer/custom-templates/components/utils';
+import { isTemplateVariablesEnabled } from '@/react/portainer/custom-templates/components/utils';
 import { TemplateTypeSelector } from '@/react/portainer/custom-templates/components/TemplateTypeSelector';
 import { AccessControlForm } from '@/react/portainer/access-control';
 import { EnvironmentId } from '@/react/portainer/environments/types';
@@ -24,6 +20,7 @@ import { FormActions } from '@@/form-components/FormActions';
 import { FormSection } from '@@/form-components/FormSection';
 
 import { EdgeTemplateSettings } from '../types';
+import { useParseTemplateOnFileChange } from '../useParseTemplateOnFileChange';
 
 import { EdgeSettingsFieldset } from './EdgeSettingsFieldset';
 import { FormValues, Method, initialBuildMethods } from './types';
@@ -55,6 +52,10 @@ export function InnerForm({
     initialValues.FileContent,
     values.FileContent,
     isEditor && !isSubmitting && !isLoading
+  );
+
+  const handleChangeFileContent = useParseTemplateOnFileChange(
+    values.Variables
   );
 
   const texts = textByType[values.Type];
@@ -180,36 +181,6 @@ export function InnerForm({
       />
     </Form>
   );
-
-  function handleChangeFileContent(value: string) {
-    setFieldValue(
-      'FileContent',
-      value,
-      isTemplateVariablesEnabled ? !value : true
-    );
-    parseTemplate(value);
-  }
-
-  function parseTemplate(value: string) {
-    if (!isTemplateVariablesEnabled || value === '') {
-      setFieldValue('Variables', []);
-      return;
-    }
-
-    const [variables, validationError] = getTemplateVariables(value);
-    const isValid = !!variables;
-
-    setFieldError(
-      'FileContent',
-      validationError ? `Template invalid: ${validationError}` : undefined
-    );
-    if (isValid) {
-      setFieldValue(
-        'Variables',
-        intersectVariables(values.Variables, variables)
-      );
-    }
-  }
 
   function handleChangeMethod(method: Method) {
     setFieldValue('FileContent', '');
